+++ idea/testData/quickfix/removeRedundantAssignment/constant.kt
@@ -1,0 +1,5 @@
+// "Remove redundant assignment" "true"
+fun test() {
+    var i: Int
+    <caret>i = 1
+}